move automata to their own module
