+++ src/Automata/rebuilding_idxA.jl
