+++ src/rewriting.jl
@@ -63,3 +63,62 @@ defined by the inverses present in alphabet `A`.
     end
     return v
 end
+
+"""
+    rewrite!(v::AbstractWord, w::AbstractWord, rws::RewritingSystem)
+Rewrite word `w` storing the result in `v` by left using rewriting rules of
+rewriting system `rws`. See [Sims, p.66]
+"""
+@inline function rewrite!(
+    v::AbstractWord,
+    w::AbstractWord,
+    rws::RewritingSystem,
+)
+    v = resize!(v, 0)
+    while !isone(w)
+        push!(v, popfirst!(w))
+        for (lhs, rhs) in rules(rws)
+            if issuffix(lhs, v)
+                prepend!(w, rhs)
+                resize!(v, length(v) - length(lhs))
+                # since suffixes of v has been already checked against rws we
+                # can break here
+                break
+            end
+        end
+    end
+    return v
+end
+
+function rewrite!(
+    v::AbstractWord,
+    w::AbstractWord,
+    idxA::Automata.IndexAutomaton{S};
+    history_tape = S[],
+) where {S}
+    resize!(history_tape, 1)
+    history_tape[1] = Automata.initial(idxA)
+
+    resize!(v, 0)
+    while !isone(w)
+        x = popfirst!(w)
+        σ = last(history_tape) # current state
+        τ = σ[x] # next state
+        @assert !isnothing(τ) "idxA doesn't seem to be complete!; $σ"
+
+        if Automata.isterminal(τ)
+            lhs, rhs = Automata.value(τ)
+            # lhs is a suffix of v·x, so we delete it from v
+            resize!(v, length(v) - length(lhs) + 1)
+            # and prepend rhs to w
+            prepend!(w, rhs)
+            # now we need to rewind the history tape
+            resize!(history_tape, length(history_tape) - length(lhs) + 1)
+            # @assert trace(v, ia) == (length(v), last(path))
+        else
+            push!(v, x)
+            push!(history_tape, τ)
+        end
+    end
+    return v
+end
